refactor(AbstractPiece): 다중 상속 요소가 없어, 추상 골격 클래스 삭제
