+++ src/main/java/janggi/domain/piece/Cha.java
@@ -9,7 +9,7 @@ import janggi.domain.status.Team;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Cha extends AbstractPiece {
+public class Cha extends Piece {
 
     private static final int SCORE = 13;
