+++ src/main/java/janggi/domain/piece/Pho.java
@@ -10,7 +10,7 @@ import janggi.domain.status.Team;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Pho extends AbstractPiece {
+public class Pho extends Piece {
 
     private static final int SCORE = 7;
